Move di to presentation
