+++ app/src/main/kotlin/ru/rambler/it/presentation/RamblerITApplication.kt
@@ -2,7 +2,7 @@ package ru.rambler.it.presentation
 
 import android.app.Application
 import io.realm.Realm
-import ru.rambler.it.other.di.initializeAppComponent
+import ru.rambler.it.presentation.internal.di.initializeAppComponent
 
 class RamblerITApplication : Application() {
