+++ app/src/main/kotlin/ru/rambler/it/presentation/internal/di/component/AppComponent.kt
@@ -1,10 +1,10 @@
-package ru.rambler.it.other.di.component
+package ru.rambler.it.presentation.internal.di.component
 
 import android.content.Context
 import dagger.Component
-import ru.rambler.it.other.di.module.AppModule
-import ru.rambler.it.other.di.module.DomainModule
-import ru.rambler.it.other.di.module.MainScreenModule
+import ru.rambler.it.presentation.internal.di.module.AppModule
+import ru.rambler.it.presentation.internal.di.module.DomainModule
+import ru.rambler.it.presentation.internal.di.module.MainScreenModule
 import javax.inject.Singleton
 
 @Singleton
