+++ app/src/main/kotlin/ru/rambler/it/presentation/internal/di/module/MainScreenModule.kt
@@ -1,10 +1,10 @@
-package ru.rambler.it.other.di.module
+package ru.rambler.it.presentation.internal.di.module
 
 import dagger.Module
 import dagger.Provides
 import ru.rambler.it.domain.interactors.EventsInteractor
 import ru.rambler.it.domain.interactors.LocalEventsInteractor
-import ru.rambler.it.other.di.scope.PerActivity
+import ru.rambler.it.presentation.internal.di.scope.PerActivity
 import ru.rambler.it.presentation.presenter.MainScreenPresenter
 import ru.rambler.it.presentation.view.interfaces.MainScreenView
